feat: Separate pages logically
