+++ src/router/routeNames.ts
@@ -3,9 +3,7 @@ export const enum RouteName {
 	Timer = 'timer',
 	Players = 'players',
 	Games = 'games',
-	WheelEffects = 'wheel-effects',
-	WheelRolls = 'wheel-rolls',
-	GameRolls = 'game-rolls',
+	Effects = 'wheel-rolls',
 	Dev = 'dev',
 	Login = 'login',
 	Signup = 'signup',
