+++ src/views/Games/GamesView.vue
@@ -1,64 +1,52 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia'
-import {
-	computed,
-	onMounted,
-	ref,
-	useTemplateRef,
-	watch,
-	watchEffect,
-} from 'vue'
-import { useFeatureGameStore } from '../../stores/feature/featureGameStore'
+import { computed, onMounted, ref, watch } from 'vue'
+import UiButton from '../../components/ui/UiButton.vue'
+import UiView from '../../components/ui/UiView.vue'
 import { LoadingState } from '../../composables/useLoading'
 import { useAuthStore } from '../../stores/authStore'
+import { useFeatureGameStore } from '../../stores/feature/featureGameStore'
+import AddGameForm from './components/AddGameForm.vue'
 
 const gameStore = useFeatureGameStore()
-const authStore = useAuthStore()
 
-const { wishlist } = storeToRefs(gameStore)
+const { current, canRoll, wishlist } = storeToRefs(gameStore)
 const showCountHint = ref(false)
-
-const addGameInput = useTemplateRef('add-game-input')
 
 gameStore.wishlistLoading.on([LoadingState.LOADED]).then(() => {
 	showCountHint.value = !gameStore.enoughGamesInWishlist
-	addGameInput.value?.focus()
 })
 
-const gameName = ref('')
+const rollText = ref('Загрузка...')
 
-const isLoading = computed(
-	() => gameStore.wishlistLoading.state === LoadingState.LOADING,
+const showWishlistCountHint = computed(
+	() =>
+		!gameStore.enoughGamesInWishlist &&
+		gameStore.currentLoading.state === LoadingState.LOADED &&
+		gameStore.wishlistLoading.state === LoadingState.LOADED,
 )
 
-/** @todo extract validation */
-const validator = computed(() => {
-	const result: { ok: boolean; message?: string } = { ok: true }
-
-	if (wishlist.value.find((u) => u.name === gameName.value)) {
-		result.ok = false
-		result.message = 'Такая игра уже есть'
-	}
-
-	return result
-})
-
-watchEffect(() => {
-	addGameInput.value?.setCustomValidity(
-		validator.value.ok ? '' : (validator.value.message ?? 'Неверные данные'),
+gameStore.currentLoading.on([LoadingState.LOADED]).then(() => {
+	watch(
+		current,
+		() => {
+			rollText.value = current.value?.name ?? 'Крути барабан'
+		},
+		{ immediate: true },
 	)
-	addGameInput.value?.reportValidity()
 })
 
-const onAddGameFormSubmit = () => {
-	if (!validator.value.ok) return
-	if (!gameName.value.length) return
+onMounted(() => {
+	if (
+		[LoadingState.ERROR, LoadingState.INIT].includes(
+			gameStore.wishlistLoading.state,
+		)
+	) {
+		gameStore.getWishlist()
+	}
+})
 
-	gameStore.addToWishlist({ name: gameName.value }).then(() => {
-		gameName.value = ''
-		showCountHint.value = !gameStore.enoughGamesInWishlist
-	})
-}
+const authStore = useAuthStore()
 
 const updateGamesOnLoginChange = () => {
 	watch(
@@ -83,44 +71,162 @@ onMounted(() => {
 </script>
 
 <template>
-	<div class="games-view">
-		<div class="container">
-			<h1 class="title">Игры</h1>
+	<UiView class="games-view">
+		<h1>Игры</h1>
 
-			<form class="add-form" @submit.prevent="onAddGameFormSubmit">
-				<input
-					class="game-input"
-					ref="add-game-input"
-					placeholder="Название игры..."
-					:disabled="isLoading"
-					v-model.trim="gameName"
-				/>
+		<div class="game-rolls">
+			<div class="roll-display">
+				<div>{{ rollText }}</div>
+			</div>
 
-				<button class="submit-button" :disabled="isLoading">Добавить</button>
-			</form>
+			<div class="status-container">
+				<p v-if="current" class="current-game">
+					Ты сейчас играешь в {{ current?.name }}.
+				</p>
 
-			<p class="hint" v-if="showCountHint">
-				Чтобы крутить следующую игру, надо 6 игр, нужно ещё
-				{{ 6 - wishlist.length }}.
-			</p>
+				<span class="status-row" v-if="current">
+					<UiButton class="status-button" @click="gameStore.finish()">
+						Закончить
+					</UiButton>
 
-			<ol class="game-list" v-if="wishlist.length">
-				<li :key="game.name" v-for="game in wishlist">
-					{{ game.name }}
-				</li>
-			</ol>
+					<UiButton class="status-button" @click="gameStore.cancel()">
+						Бросить
+					</UiButton>
+				</span>
+			</div>
 
-			<p class="message" v-else-if="isLoading && !gameName.length">
-				Загружаем игры...
-			</p>
-
-			<p class="message" v-else>У тебя ещё нет игр!</p>
+			<div class="action-row">
+				<UiButton
+					class="roll-button"
+					:disabled="!canRoll"
+					@click="gameStore.roll()"
+				>
+					Прокрутить
+				</UiButton>
+			</div>
 		</div>
-	</div>
+
+		<div class="games-list">
+			<div class="container">
+				<AddGameForm />
+
+				<p class="hint" v-if="showCountHint">
+					Чтобы крутить следующую игру, надо 6 игр, нужно ещё
+					{{ 6 - wishlist.length }}.
+				</p>
+
+				<ol class="game-list" v-if="wishlist.length">
+					<li :key="game.name" v-for="game in wishlist">
+						{{ game.name }}
+					</li>
+				</ol>
+
+				<p class="message" v-else>У тебя ещё нет игр!</p>
+			</div>
+		</div>
+	</UiView>
 </template>
 
 <style scoped>
-.games-view {
+.games-view > * {
+	display: flex;
+	flex-direction: column;
+	width: 100%;
+}
+
+.game-rolls {
+	display: flex;
+	flex-direction: column;
+	height: 100%;
+	inline-size: 100%;
+	place-content: center;
+	gap: 48px;
+	padding-bottom: 80px;
+}
+
+.effects-link {
+	place-self: center;
+	color: #3b82f6;
+}
+
+.effects-link:hover {
+	text-decoration: underline;
+}
+
+.title {
+	text-align: center;
+	font-size: 2.25rem;
+	font-weight: 600;
+}
+
+.roll-display {
+	display: flex;
+	height: 160px;
+	width: 100%;
+	align-items: center;
+	justify-content: center;
+	gap: 16px;
+	border: 1px solid #e2e8f0;
+}
+
+.status-container {
+	border-radius: 6px;
+	background-color: #e2e8f0;
+	padding: 16px 32px;
+	width: 100%;
+}
+
+.status-container:empty {
+	display: none;
+}
+
+.status-row {
+	display: flex;
+	gap: 8px;
+}
+
+.current-game {
+	display: flex;
+	width: 100%;
+	align-items: center;
+	gap: 16px;
+}
+
+.status-button {
+	height: 72px;
+}
+
+.info-item {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 12px;
+	color: #0f172a;
+}
+
+.add-games-link {
+	box-sizing: border-box;
+	border: 2px solid #3b82f6;
+	padding: 8px 16px;
+	color: #3b82f6;
+}
+
+.add-games-link:hover {
+	text-decoration: underline;
+}
+
+.roll-button {
+	height: 60px;
+	font-size: 1.25rem;
+	width: 240px;
+}
+
+.action-row {
+	display: flex;
+	justify-content: center;
+}
+
+.games-list {
 	display: flex;
 	width: 100%;
 	flex-direction: column;
@@ -130,7 +236,7 @@ onMounted(() => {
 
 .container {
 	display: flex;
-	width: 720px;
+	width: 100%;
 	flex-direction: column;
 	gap: 32px;
 }
@@ -184,3 +290,5 @@ onMounted(() => {
 	width: 100%;
 }
 </style>
+
+<script setup lang="ts"></script>
