+++ src/views/Players/PlayersView.vue
@@ -1,3 +1,21 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import UiView from '../../components/ui/UiView.vue'
+</script>
 
-<template>В работе</template>
+<template>
+	<UiView>
+		<h1>Игроки</h1>
+
+		<p>В работе...</p>
+	</UiView>
+</template>
+
+<style scoped>
+h1 {
+	margin-block-end: 1rem;
+}
+
+p {
+	margin-block-end: 10rem;
+}
+</style>
